requirements.txt
pip

aiohttp
cachetools
mysql-connector-python
nextcord[speed]
nextcord[voice]
PyNaCl
requests
youtube_dl
youtube_search